[UPDATE] border's logic for optimizing

+++ application/sources/app/game/game_zomwar/zw_game_bang.h
@@ -24,7 +24,7 @@
 
 typedef struct
 {
-    uint16_t x, y;
+    uint8_t x, y;
     uint8_t action_image;
     bool visible;
 } zw_game_bang_t;
